+++ NEWS.md
@@ -1,7 +1,11 @@
 ## exprso 0.1.7.9000
 ---------------------
-* Fixed `arrayExprs` error when reading directly from file
-* `pl` functions now store some run parameters in results
+* Improvements to data import
+  * Fixed `arrayExprs` error when reading directly from file
+* Improvements to `pl` functions
+  * `pl` functions now store run parameters in results
+  * Embedding `plMonteCarlo` now calculates outer-fold accuracy
+  * Embedding `plNested` now calculates outer-fold accuracy
 * Added F1000Research citation to `inst/CITATION`
 
 ## exprso 0.1.7