+++ NEWS.md
@@ -2,6 +2,8 @@
 ---------------------
 * Improvements to `arrayExprs`
   * Fixed an import error when reading directly from file
+* Improvements to `calcStats`
+  * All `NaN` and `NA` performance metrics now replaced with 0
 * Improvements to `pl` functions
   * `pl` functions now store run parameters in results
   * Embedding `plMonteCarlo` now calculates outer-fold accuracy